Refactor controllers and introduce cart management features
Updated `AdminsController`, `AuthsController`, and `ProductsController` to inherit from `ApiControllerBase` for standardization. Added `CartDTO` for cart item management and introduced `CartService`, `Cart`, and `CartItem` entities. Updated `ProductDTO` to use `decimal` for price values and created new repositories for cart management. Added migration for new database tables and improved timestamp handling in `RepositoryBase`.

+++ VTT_SHOP_CORE/DTOs/ProductDTO.cs
@@ -7,7 +7,7 @@
         public string SlugName { get; set; } = string.Empty ;
         public string Description { get; set; } = string.Empty ;
         public int Quantity { get; set; } 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
     }
 
     public class CreateProductDTO
@@ -15,7 +15,7 @@
         public string Name { get; set; }
         public string Description { get; set; } = string.Empty;
         public int Quantity { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
     }
 
     public class UpdateProductDTO
@@ -24,6 +24,6 @@
         public string Name { get; set; } = string.Empty;
         public string? Description { get; set; }
         public int Quantity { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
